Fix. Seed value now parsed as a Long.
Previously, the seed value was being parsed as an integer. This worked fine for
trivial cases (such as 1234) but failed when current timestamps were used, since
the timestamp value is larger than that permitted by integers.

Seed value is now parsed as a Long.

+++ TrailMakingTest.pde
@@ -555,7 +555,7 @@ long loadRandomSeed() {
 
   long seed = defaultValue;
   try {
-    seed = Integer.parseInt(seedString);
+    seed = Long.parseLong(seedString);
   } catch (Exception e) {
     return defaultValue;
   }
